refactor(aircraft): consolidate byte-identical helpers into BaseAircraftDefinition
