+++ MSFSBlindAssist/Aircraft/FenixA320Definition.cs
@@ -12666,30 +12666,8 @@ public class FenixA320Definition : BaseAircraftDefinition
         }
     }
 
-    private void RequestFuelQuantity(SimConnect.SimConnectManager simConnectMgr)
-    {
-        var simConnect = simConnectMgr.SimConnectInstance;
-        if (simConnectMgr.IsConnected && simConnect != null)
-        {
-            try
-            {
-                var tempDefId = SimConnect.SimConnectManager.DATA_DEFINITIONS.DEF_FUEL_QUANTITY;
-                simConnect.ClearDataDefinition(tempDefId);
-                simConnect.AddToDataDefinition(tempDefId,
-                    "FUEL TOTAL QUANTITY WEIGHT", "pounds",
-                    Microsoft.FlightSimulator.SimConnect.SIMCONNECT_DATATYPE.FLOAT64, 0.0f, 0);
-                simConnect.RegisterDataDefineStruct<SimConnect.SimConnectManager.SingleValue>(tempDefId);
-                simConnect.RequestDataOnSimObject(SimConnect.SimConnectManager.DATA_REQUESTS.REQUEST_FUEL_QUANTITY,
-                    tempDefId, Microsoft.FlightSimulator.SimConnect.SimConnect.SIMCONNECT_OBJECT_ID_USER,
-                    Microsoft.FlightSimulator.SimConnect.SIMCONNECT_PERIOD.ONCE,
-                    Microsoft.FlightSimulator.SimConnect.SIMCONNECT_DATA_REQUEST_FLAG.DEFAULT, 0, 0, 0);
-            }
-            catch (Exception ex)
-            {
-                Log.Debug("Fenix", $"Error requesting fuel quantity: {ex.Message}");
-            }
-        }
-    }
+    // RequestFuelQuantity moved to BaseAircraftDefinition (byte-identical Fenix/FBW A320
+    // pair, now parameterized on the Log.Debug category).
 
     /// <summary>
     /// Delay (ms) after the initial reset-to-0 before the button is pressed (0 → 1).
@@ -13331,7 +13309,7 @@ public class FenixA320Definition : BaseAircraftDefinition
                 return true;
 
             case HotkeyAction.ReadFuelQuantity:
-                RequestFuelQuantity(simConnect);
+                RequestFuelQuantity(simConnect, "Fenix");
                 return true;
 
             // Fenix has no waypoint info feature — W key repurposed for gross weight (lbs)
